Lock down user, post and comment mutations

+++ src/resolvers/Mutation.js
@@ -23,13 +23,15 @@ export default {
         };
     },
     async deleteUser(parent, args, { prisma }, info) {
+        const userId = getUserId(request);
 
-        return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+        return prisma.mutation.deleteUser({ where: { id: userId } }, info);
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
 
         return prisma.mutation.updateUser({
-            where: { id: args.id },
+            where: { id: userId },
             data: args.data
         }, info);
     },
@@ -69,24 +71,50 @@ export default {
             } 
         }, info);
     },
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists) {
+            throw new Error('Unable to delete post.');
+        }
         
         return prisma.mutation.deletePost({ where: { id: args.id } }, info);
     },
-    async updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!postExists) {
+            throw new Error('Unable to update post.');
+        }
         
         return prisma.mutation.updatePost({
             where: { id: args.id },
             data: args.data
         }, info);
     },
-    async createComment(parent, args, { prisma }, info) {
+    async createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -97,11 +125,35 @@ export default {
             }
         }, info);
     },
-    async deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!commentExists) {
+            throw new Error('Unable to delete comment.');
+        }
 
         return prisma.mutation.deleteComment({ where: { id: args.id } }, info);
     },
-    async updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request);
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        });
+
+        if (!commentExists) {
+            throw new Error('Unable to update comment.');
+        }
 
         return prisma.mutation.updateComment({
             where: { id: args.id },
